Added better naviagtion so you can go through pretty much the whole app Need to refactor becuase of some duplicate code in the change state Method Some CSS issues on the bottom nav bars Gotta talk about color themes for the app
